+++ Pipfile
@@ -5,7 +5,6 @@ name = "pypi"
 
 [packages]
 requests-html = "*"
-datafiles = "*"
 scikit-learn = "*"
 pandas = "*"
 joblib = "*"
