+++ get_graph_data.py
@@ -25,12 +25,23 @@ def to_graphql_data():
                 with open('data/'+itemdir, 'r') as f:
                     item = json.load(f)
                     url_set.add(item['url'])
-                    data['nodes'].append({
+                    d = {
                         'id': item['url'],
-                        'label': item['url'],
+                        'label': item['name'],  # 显示站点名称
                         'title': item['name'],
-                        'group': item['url'].split(".")[-1]
-                    })
+                        'group': item['url'].split(".")[-1],
+                        # 'brokenImage':  "https://gine.me/icons/icon-48x48.png",
+                    }
+
+                    # fixme 默认加载的站点数量较多，显示 icon 会有大量的网络请求。
+                    # if item['icon']:
+                    #     d = {
+                    #         **d,
+                    #         'shape': "circularImage",
+                    #         'image':  item['icon']
+                    #     }
+
+                    data['nodes'].append(d)
                     url_friends_map[item['url']] = item['friends']
             except Exception as e:
                 print(e)
