+++ site/src/components/SiteCard.js
@@ -1,4 +1,4 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import CardActionArea from '@material-ui/core/CardActionArea';
@@ -26,8 +26,7 @@ const useStyles = makeStyles({
 export default function MediaCard() {
     const classes = useStyles();
     const { state, dispatch } = useContext(AppContext)
-    const { hideCard, setHideCard } = useState(false)
-    const { selectedSite, depath } = state
+    const { selectedSite } = state
     return (
         <>
             {selectedSite && <Card className={classes.card}>
